requirements.txt
ccxt>=4.0.0
numpy>=1.21.0
python-dotenv>=0.19.0
